% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Sanity check of input data object}
\usage{
check_data(sce.object)
}
\arguments{
\item{sce.object}{A \code{FRmatch} input data object. See details in \code{\link[FRmatch]{sce.example}}.}
}
\value{
A data object that passes this sanity check if no error occurs.
}
\description{
For input data, we choose to summarize the gene expression data and metadata needed for \code{\link[FRmatch]{FRmatch}} in a data object of
the \code{\link[SingleCellExperiment]{SingleCellExperiment}} class, which is a light-weighted container for single-cell genomics data.
This function helps to check if users' data objects can be used in the \code{FRmatch} function.
}
\details{
For the FR-Match method, the following data are essential:
\itemize{
\item a gene expression data matrix (gene by cell)
\item cell cluster membership information
\item NS-Forest (or your own) marker genes for the reference dataset
}
In addition, information such as F-meansure and cluster order are not essential, but will facilitate visualization and customized analysis
provided in this package.
}
\seealso{
The \link[SingleCellExperiment]{SingleCellExperiment} data class.
}
