% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_object.R
\name{check_data_object}
\alias{check_data_object}
\title{Check data object}
\usage{
check_data_object(sce.object, is.ref = TRUE, verbose = TRUE)
}
\arguments{
\item{sce.object}{Data object of the \link[SingleCellExperiment]{SingleCellExperiment} class.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{verbose}{Boolean variable indicating if to print details to the console. Default: \code{TRUE}.}
}
\value{
A data object that passes this sanity check if no error occurs.
}
\description{
We choose to summarize the clustered gene expression data and metadata in a data object of
the \code{\link[SingleCellExperiment]{SingleCellExperiment}} class.
This function helps to check if the data object is compatible with \code{\link[FRmatch]{FRmatch}}.
}
\details{
For the FR-Match algorithm, the following data elements are essential:
\itemize{
\item gene expression data matrix
\item cell cluster membership
\item \emph{informative} marker genes of the reference dataset
}
In addition, metadata such as \code{cluster_marker_info, f_score, cluster_order} are not essential,
but will facilitate visualization and more customized analyses provided in this package.
}
\seealso{
The \link[SingleCellExperiment]{SingleCellExperiment} class.
}
