% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dropout.R
\name{check_dropout}
\alias{check_dropout}
\title{Check per-cluster dropout rate for cluster marker genes}
\usage{
check_dropout(sce.object, return.value = TRUE, plot.dropout = FALSE,
  cellwidth = 15, cellheight = 10, ...)
}
\arguments{
\item{sce.object}{A \code{SingleCellExperiment} object filled with necessary information for \code{FRmatch}.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{return.value}{Logical variable indicating if to return the dropout rate values to R console. Default: \code{TRUE}.}

\item{plot.dropout}{Logical variable indicating if to return a plot of dropout rates. Default: \code{FALSE}.}

\item{cellwidth, cellheight, ...}{Additional plotting parameters passed to \code{\link[pheatmap]{pheatmap}}.}
}
\value{
A numeric matrix of dropout rates per cluster per marker gene, and optionally, a plot.
}
\description{
A function that calculates and plots the dropout rate for a \code{SingleCellExperiment} object
customized with cluster and NS-Forest marker gene information.
}
\examples{
\dontrun{
data("sce.example")
check_dropout(sce.example, plot.dropout=TRUE, main="Dropout rate plot")
}

}
\seealso{
The \link[SingleCellExperiment]{SingleCellExperiment} class.
}
\author{
Yun Zhang, \email{zhangy@jcvi.org}; Brian Aevermann, \email{baeverma@jcvi.org}; Richard Scheuermann, \email{RScheuermann@jcvi.org}.
}
