% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FR.test.subsamp.R
\name{FR.test.subsamp.each}
\alias{FR.test.subsamp.each}
\title{Friedman-Rafsky (FR) test with subsampling}
\usage{
FR.test.subsamp.each(samp1, samp2, subsamp.size, subsamp.iter, ...)
}
\arguments{
\item{samp1}{Numeric matrix or data frame for Sample 1.}

\item{samp2}{Numeric matrix or data frame for Sample 2.}

\item{subsamp.size, subsamp.iter}{Cluster size and number of iterations for subsampling.}

\item{...}{Additional parammeters passed to \code{\link[FRmatch]{FR.test}}.}
}
\value{
Median p-value from the subsampling itertations and its corresponding test statistics will be returned.
\item{runs}{Total number of subtrees.}
\item{runs.samp1}{Number of subtrees of Sample 1.}
\item{runs.samp2}{Number of subtrees of Sample 2.}
\item{stat}{The standardized FR statistic.}
\item{p.value}{P-value of the FR test.}
}
\description{
FR test with subsampling, which is provided for comparing two samples with one or both large sample size(s).
}
\examples{
\dontrun{
samp1 <- matrix(rnorm(1000),nrow=5)
samp2 <- matrix(rnorm(1000),nrow=5)
FR.test.subsamp.each(samp1, samp2)
}
}
