% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRmatch.R
\name{FRmatch}
\alias{FRmatch}
\title{FR-Match for single cell RNA-seq data}
\usage{
FRmatch(
  sce.query,
  sce.ref,
  use.cosine = TRUE,
  filter.size = 5,
  filter.fscore = NULL,
  filter.nomarker = FALSE,
  add.pseudo.marker = FALSE,
  pseudo.expr = 1,
  subsamp.size = 10,
  subsamp.iter = 1000,
  subsamp.seed = 1,
  numCores = NULL,
  prefix = c("query.", "ref."),
  verbose = 1,
  return.all = FALSE,
  ...
)
}
\arguments{
\item{sce.query}{Data object of the \link[SingleCellExperiment]{SingleCellExperiment} class for query experiment.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{sce.ref}{Data object of the \link[SingleCellExperiment]{SingleCellExperiment} class for reference experiment.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{use.cosine}{Boolean variable indicating if to use cosine distance in \code{\link[FRmatch]{FR-test}}. Default: \code{TRUE}.}

\item{filter.size, filter.fscore, filter.nomarker}{Filtering out small/poor-quality/no-marker clusters. Default: \code{filter.size=5}, filter based on the number
of cells per cluster; \code{filter.fscore=NULL}, do not filter based on the F-beta score, otherwise specify a numeric value between 0 and 1;
\code{filter.nomarker=FALSE}, filter based on the boolean variable indicating if to filter reference clusters with no marker genes available in query.}

\item{add.pseudo.marker, pseudo.expr}{Adding pseudo marker to stabilize no expression clusters in the marker gene feature space.
Default: \code{add.pseudo.marker=FALSE}, boolean. Pseudo marker expression values are drawn from uniform distribution from 0 to \code{pseudo.expr}.
Default: \code{pseudo.expr=1}, numeric, for the min-max scaled data after normalization.}

\item{subsamp.size, subsamp.iter, subsamp.seed}{Numeric variables for iterative subsampling size, number of iterations, and random seed for iterations. YMMV.}

\item{numCores}{Number of cores for parallel computing. Default: \code{NULL}, use the maximum number of cores detected by \code{\link[parallel]{detectCores}}.
Otherwise, specify by an integer value.}

\item{prefix}{Prefix names for query and reference clusters. Default: \code{prefix=c("query.", "ref.")}.}

\item{verbose}{Numeric variable indicating levels of details to be printed. Default: \code{1}, only print major steps.
If \code{0}, no verbose; if \code{2}, print all, including warnings.}

\item{return.all}{Boolean variable indicating if to return all results (such as runs, etc.). Default: \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[FRmatch]{FRtest}}, including \code{use.cosine}.}
}
\value{
A list of:
\item{settings}{Record of customized parameter settings specified in the function.}
\item{pmat}{A matrix of raw p-values. Rows are reference clusters, and columns are query clusters.}
\item{statmat}{A matrix of FR statistics. Rows are reference clusters, and columns are query clusters.}
If \code{return.all = TRUE}, more intermediate results are returned.
}
\description{
This is a user-end wrapper function that implements the steps of cell type matching between two single cell RNA-seq experiments
(namely, \code{query} and \code{reference}) using the cluster-to-cluster FR-Match approach.
}
\examples{
\dontrun{
data("sce.example")
FRmatch(sce.example, sce.example)
}

}
\seealso{
Visualization of matching results using \code{\link[FRmatch]{plot_FRmatch}}, \code{\link[FRmatch]{plot_bi_FRmatch}}.
}
