% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRmatch_cell2cluster.R
\name{FRmatch_cell2cluster}
\alias{FRmatch_cell2cluster}
\title{FR-Match cell-to-cluster matching}
\usage{
FRmatch_cell2cluster(
  sce.query,
  sce.ref,
  use.cosine = TRUE,
  feature.selection = "reference.markers",
  filter.size = 5,
  filter.fscore = NULL,
  filter.nomarker = FALSE,
  add.pseudo.marker = FALSE,
  pseudo.expr = 1,
  subsamp.size = 10,
  subsamp.iter = 2000,
  subsamp.seed = 1,
  subsamp.iter.custom = FALSE,
  subsamp.iter.custom.k = 5,
  numCores = NULL,
  prefix = c("query.", "ref."),
  verbose = 1,
  ...
)
}
\arguments{
\item{sce.query}{Data object of the \link[SingleCellExperiment]{SingleCellExperiment} class for query experiment.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{sce.ref}{Data object of the \link[SingleCellExperiment]{SingleCellExperiment} class for reference experiment.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{use.cosine}{Boolean variable indicating if to use cosine distance in \code{\link[FRmatch]{FR-test}}. Default: \code{TRUE}.}

\item{feature.selection}{Which set of features to use for the matching space? Default: \code{feature.selection="reference.markers"}, use reference marker genes.
If \code{feature.selection="query.genes"}, use all query genes as the feature space, e.g. the query genes are probe genes for spatial transcriptomics experiment.}

\item{filter.size, filter.fscore, filter.nomarker}{Filtering out small/poor-quality/no-marker clusters. Default: \code{filter.size=5}, filter based on the number
of cells per cluster; \code{filter.fscore=NULL}, do not filter based on the F-beta score, otherwise specify a numeric value between 0 and 1;
\code{filter.nomarker=FALSE}, filter based on the boolean variable indicating if to filter reference clusters with no marker genes available in query
in the case \code{feature.selection="reference.markers"}.}

\item{add.pseudo.marker, pseudo.expr}{Adding pseudo marker to stabilize no expression clusters in the marker gene feature space.
Default: \code{add.pseudo.marker=FALSE}, boolean. Pseudo marker expression values are drawn from uniform distribution from 0 to \code{pseudo.expr}.
Default: \code{pseudo.expr=1}, numeric, for the min-max scaled data after normalization.}

\item{subsamp.size, subsamp.iter, subsamp.seed}{Numeric variables for iterative subsampling size, number of iterations, and random seed for iterations. YMMV.}

\item{subsamp.iter.custom, subsamp.iter.custom.k}{Customization of number of iterations when query cluster sizes vary a lot. Default: \code{FALSE}, no customization.
If \code{TRUE}, set a numeric value to \code{subsamp.iter.custom.k}, e.g. 5 (default), which will set the number of iterations to be at least \code{subsamp.iter}
or 5 times the cluster size for each query cluster.}

\item{numCores}{Number of cores for parallel computing. Default: \code{NULL}, use the maximum number of cores detected by \code{\link[parallel]{detectCores}}.
Otherwise, specify by an integer value.}

\item{prefix}{Prefix names for query and reference clusters. Default: \code{prefix=c("query.", "ref.")}.}

\item{verbose}{Numeric value indicating levels of details to be printed. Default: \code{1}, only print major steps.
If \code{0}, no verbose; if \code{2}, print all, including warnings.}

\item{...}{Additional arguments passed to \code{\link[FRmatch]{FRtest}}, including \code{use.cosine}.}
}
\value{
A list of:
\item{settings}{Record of customized parameter settings specified in the function.}
\item{pmat}{A cell-by-cluster (a.k.a. query cell by reference cluster) matrix of raw p-values.}
\item{cell2cluster}{A data frame of cell-to-cluster matches summarized from the \code{pmat}.}
Columns in \code{cell2cluster} are:
\item{query.cell}{Query cell ID.}
\item{query.cluster}{Original cluster of query cell.}
\item{match}{Matched reference cluster for the query cell.}
\item{score}{Confidence score for \code{match}, which is the maximum value of the corresponding query cell row in \code{pmat}.}
}
\description{
This is a user-end wrapper function that implements the steps of cell type matching between two single cell RNA-seq experiments
(namely, \code{query} and \code{reference}) using the cell-to-cluster FR-Match approach that matches each query cell to a reference cluster.
}
\details{
This implementation is \code{FRmatch} with an iterative subsampling scheme, which is a bootstrap-like approach to randomly select a smaller
set of cells from a query cluster and quantify the confidence score of the selected cells belonging to certain reference cell type
using the p-value outputted from \code{FRtest} (i.e. a larger p-value indicates higher probability of a match, and vice versa).

This function assigns the cluster-level p-value to each selected query cell, and updates the assigned p-value if the query cell is reselected
from the iterative procedure and assigned a higher p-value. The output from this implementation includes a cell-by-cluster
(a.k.a. query cell by reference cluster) matrix of p-values.
}
\examples{
\dontrun{
data("sce.example")
FRmatch_cell2cluster(sce.example, sce.example)
}

}
\seealso{
Visualization of matching results using \code{\link[FRmatch]{plot_FRmatch_cell2cluster}}.
}
