% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{get_values}
\alias{get_values}
\title{Utility function to get the plotted values}
\usage{
get_values(output)
}
\arguments{
\item{output}{The output from \code{plot_FRmatch()} or \code{plot_FRmatch()} if \code{return.value=T}.}
}
\value{
A data frame of row names, column names, and values.
}
\description{
Get the plotted values from an output matrix if \code{return.value = T}.
}
\seealso{
\code{\link[FRmatch]{plot_FRmatch}, \link[FRmatch]{plot_bi_FRmatch}}.
}
