% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_dropout.R
\name{impute_dropout}
\alias{impute_dropout}
\title{Imputation for dropouts}
\usage{
impute_dropout(sce.object)
}
\arguments{
\item{sce.object}{A \code{SingleCellExperiment} object filled with necessary information for \code{FRmatch}.
See details in \code{\link[FRmatch]{sce.example}}.}
}
\value{
A \code{SingleCellExperiment} object with imputed data replaced in the \code{logcounts} assay of the original object.
}
\description{
Impute the reference dataset only. Do NOT impute the query dataset. See 'Details' for imputation scheme.
}
\details{
Imputation scheme for NS-Forest markers:
\itemize{
\item Assume that the cell-type-specific NS-Forest markers have non-zero expression in their corresponding cell type clusters.
\item For each marker gene, impute the dropout values \emph{only} in the cluster that it marks.
\item Single imputation by randomly drawing from an empirical distribution defined by the non-dropouts in that cluster.
}
}
\examples{
\dontrun{
data("sce.example")
check_dropout(sce.example, plot.dropout=TRUE, main="Dropout rate before imputation")
sce.example.imputed <- impute_dropout(sce.example)
check_dropout(sce.example.imputed, plot.dropout=TRUE, main="Dropout rete after imputation")
}
}
\seealso{
The \link[SingleCellExperiment]{SingleCellExperiment} class.
}
\author{
Yun Zhang, \email{zhangy@jcvi.org}; Brian Aevermann, \email{baeverma@jcvi.org}; Richard Scheuermann, \email{RScheuermann@jcvi.org}.
}
