% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{Normalization function for FR-Match pipeline}
\usage{
normalization(sce.object, scale = TRUE, norm.by = NULL)
}
\arguments{
\item{sce.object}{Data object.}

\item{scale}{Boolean variable indicating if to scale the per gene expression range to 0 and 1. Default: \code{TRUE}.}

\item{norm.by}{If \code{norm.by=NULL} (default), do nothing. If \code{norm.by="median"}, to weight each gene by its median then rescale to unit range.
If \code{norm.by="mean"}, to weight each gene by its mean then rescale to unit range.}
}
\value{
A data object with normalized expression data.
}
\description{
Normalization function for FR-Match pipeline
}
