% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bilateral_FRmatch.R
\name{plot_bilateral_FRmatch}
\alias{plot_bilateral_FRmatch}
\title{Visualization of bilateral matching restuls of FRmatch}
\usage{
plot_bilateral_FRmatch(rst.FRmatch.E1toE2, rst.FRmatch.E2toE1,
  prefix.E1 = "E1", prefix.E2 = "E2", p.adj.method = "BY",
  sig.level = 0.05, reorder = TRUE, return.value = FALSE,
  cellwidth = 10, cellheight = 10, main = NULL, ...)
}
\arguments{
\item{rst.FRmatch.E1toE2, rst.FRmatch.E2toE1}{Retults outputed from the \code{\link[FRmath]{FRmatch}} function.}

\item{prefix.E1, prefix.E2}{Prefix names for E1 and E2. Default: \code{"E1"} and \code{"E2"}, respectively.}

\item{p.adj.method}{P-value adjustment method for multiple comparison correction. Default: \code{"BY"}.
Please see \code{\link[stats]{p.adjust.methods}}.}

\item{sig.level}{Numeric variable that sets the significance level for the adjusted p-values, above which is a match.
Default value: \code{0.05}.}

\item{reorder}{Logical variable indicating if to reorder the columns of heatmap for better interpretability
(aligning matches in the diagonal). Default: \code{TRUE}.}

\item{return.value}{Logical variable indicating if to return the values correponding to the plot. Default: \code{FALSE}.}

\item{cellwidth, cellheight, main, ...}{Additional plotting parameters passed to \code{\link[pheatmap]{pheatmap}}.}
}
\value{
Optionally, a numeric matrix corresponding to the values on the plot.
}
\description{
This function combines two reciprocal \code{FRmatch} result objects (i.e. experiment 1 (hereinafter, E1) as query
mapping to E2 as reference, and E2 as query mapping to E1 as reference), and plots the matches that are identified
in both ways or just in one way or no match.
}
\seealso{
\link[FRmatch]{plot_FRmatch}.
}
