% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FRmatch_cell2cluster.R
\name{plot_FRmatch_cell2cluster}
\alias{plot_FRmatch_cell2cluster}
\title{Plotting function for FR-Match cell2cluster results}
\usage{
plot_FRmatch_cell2cluster(
  rst.cell2cluster,
  type = "match.prop",
  p.adj.method = "BH",
  sig.level = 0.05,
  reorder = TRUE,
  return.value = FALSE,
  filename = NA,
  width = NULL,
  height = NULL
)
}
\description{
Plotting function for FR-Match cell2cluster results
}
