% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_most_similar_cluster.R
\name{predict_most_similar_cluster}
\alias{predict_most_similar_cluster}
\title{Prediction function for FR-Match results}
\usage{
predict_most_similar_cluster(rst.FRmatch, p.adj.method = "BY")
}
\arguments{
\item{rst.FRmatch}{The \code{\link[FRmatch]{FRmatch}} output.}

\item{p.adj.method}{P-value adjustment method for multiple hypothesis testing correction. Default: \code{"BY"}.
For more options, please see \code{\link[stats]{p.adjust.methods}}.}
}
\value{
A data frame of the most similar reference cluster for each query cluster,
which is the "match" with the highest adjusted p-value. If there are tied p-values, all will be returned.
}
\description{
This function takes in the \code{\link[FRmatch]{FRmatch}} output and predicts the most similar reference cluster for each query cluster.
}
