% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce.example}
\alias{sce.example}
\title{FRmatch input data}
\format{Please follow the nomenclature and data format of this example.
\describe{
\item{\code{metadata}}{List of 3.
\itemize{
  \item \code{cluster_marker_info} are the cluster-specific NS-Forest marker genes.
  \item \code{fsocres} are F-measures associated with the NS-Forest marker genes for each cluster.
  \item \code{cluster_order} is some specific order of the clusters to be used for plotting.
}}
\item{\code{assays}}{Data assay \code{logcounts} is the expression data matrix (gene by cell).}
\item{\code{rownames}}{Gene names.}
\item{\code{rowData}}{Column \code{NSF_markers} is a logical vector indicating if a gene is an NS-Forest marker gene.}
\item{\code{colnames}}{Cell IDs.}
\item{\code{colData}}{Column \code{cluster_membership} is a character vector of cell cluster membership.}
}}
\usage{
sce.example
}
\description{
This page describes the \link[SingleCellExperiment]{SingleCellExperiment} data object needed for the \code{\link[FRmatch]{FRmatch}}
and other functions in this package.
DESCRIPTION OF MTG DATA.
This example dataset contain 865 cells with 16497 genes.
}
\seealso{
The \link[SingleCellExperiment]{SingleCellExperiment} data class.
}
\keyword{datasets}
