% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Sanity check for input data object}
\usage{
check_data(sce.object)
}
\arguments{
\item{sce.object}{A \code{SingleCellExperiment} object filled with necessary information for \code{FRmatch}.
See details in \code{\link[FRmatch]{sce.example}}.}
}
\value{
A \code{SingleCellExperiment} object that passes this sanity check if no error occurs.
}
\description{
For input data, we choose to summarize the gene expression data and metadata needed for \code{FRmatch} in
the \code{\link[SingleCellExperiment]{SingleCellExperiment}} class, which is a light-weighted container for single-cell genomics data.
This function helps to check if users' data objects fit the \code{FRmatch} context.
}
\details{
For FRmatch, the following data items are essential:
\itemize{
\item a gene expression data matrix (gene by cell)
\item cluster membership information for each cell
\item NS-Forest (or your own) marker genes for the reference clusters
}
Additional items, such as F-meansure (i.e. a quality score associated with NS-Forest markers) and preferred cluster order,
are not essential, but will facilitate graphical tools provided in this package.
}
\seealso{
The \link[SingleCellExperiment]{SingleCellExperiment} class.
}
