% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRmatch.R
\name{FRmatch}
\alias{FRmatch}
\title{Cross-comparison of two single-cell RNA-seq experiments using Friedman-Rafsky test}
\usage{
FRmatch(sce.query, sce.ref, method = "none", imputation = FALSE,
  numCores = NULL, return.all = FALSE, ...)
}
\arguments{
\item{sce.query}{Query experiment object of \code{SingleCellExperiment} class with necessary information.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{sce.ref}{Reference experiment object of \code{SingleCellExperiment} class with necessary information.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{method}{RIGHT NOW ONLY IMPLEMENTED THE STANDARD FR TEST.}

\item{imputation}{Logical variable indicating if to impute dropout values in the reference experiment. Default: \code{FALSE}.
See details in \code{\link[FRmatch]{impute_dropout}}.}

\item{numCores}{\code{NULL} or an integer that specifies the number of cores to be used in parallel computing.
Default: \code{NULL}, which sets \code{numCores = detectCores()}. See more in \code{\link[parallel]{detectCores}}.}

\item{return.all}{Logical variable indicating if to return all results (such as number of subtrees, etc.).
Default: \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[FRmatch]{FR.test}}.}
}
\value{
A list of FR test results:
\item{statmat}{A matrix of all FR statistics for pairwise comparison. Rows are reference clusters, and columns are query clusters.}
\item{pmat}{A matrix of all p-values for pairwise comparison. Rows are reference clusters, and columns are query clusters.}
If \code{return.all = TRUE}, more results are returned in \code{all.restuls} element of the list.
}
\description{
This is a user-end function that wraps the steps to cross-compare two single-cell RNA-seq experiments (namely, query and reference)
with cluster information and NS-Forest marker genes using Friedman-Rafsky (FR) test based methods.
This function inputs two objects of \link[SingleCellExperiment]{SingleCellExperiment} class with necessary information,
and returns FR statistics (a.k.a. scores) and p-values of the FR test.
}
\seealso{
The \link[SingleCellExperiment]{SingleCellExperiment} class.
}
\author{
Yun Zhang, \email{zhangy@jcvi.org}; Brian Aevermann, \email{baeverma@jcvi.org}; Richard Scheuermann, \email{RScheuermann@jcvi.org}.
}
