% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRmatch.R
\name{FRmatch}
\alias{FRmatch}
\title{Cell type cluster matching of single-cell RNA-seq data using Friedman-Rafsky test}
\usage{
FRmatch(sce.query, sce.ref, imputation = FALSE, filter.size = 10,
  filter.fscore = NULL, method = "subsampling", subsamp.size = 10,
  subsamp.iter = 1001, subsamp.seed = 1, numCores = NULL,
  return.all = FALSE, ...)
}
\arguments{
\item{sce.query}{Data object of the \link[SingleCellExperiment]{SingleCellExperiment} data class for query experiment.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{sce.ref}{A Data object of the \link[SingleCellExperiment]{SingleCellExperiment} data class for reference experiment.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{imputation}{Logical variable indicating if to impute expression zero values for the reference experiment. Default: \code{FALSE}.
See details in \code{\link[FRmatch]{impute_dropout}}.}

\item{filter.size, filter.fscore}{Criteria for filtering small/poor clusters. Default: \code{filter.size=10}, filter based on the number
of cells per cluster; \code{filter.fscore=NULL}, filter based on the f-measure associated with NS-Forest markers (\code{NULL} or numeric value).}

\item{method}{Adapted methods of FR test. Default: \code{method="subsampling"} is to apply equal size subsampling of the comparing clusters for
the FR test. Option: \code{method="none"} is the standard FR test without modification.}

\item{subsamp.size, subsamp.iter, subsamp.seed}{Cluster size, number of iterations, and random seed for subsampling.
Default: \code{10, 1001, 1}, respectively.}

\item{numCores}{\code{NULL} or an integer that specifies the number of cores for parallel computing.
Default: \code{NULL}, is to set the maximum number of cores detected by \code{\link[parallel]{detectCores}} if not specified.}

\item{return.all}{Logical variable indicating if to return all results (such as number of runs, etc.). Default: \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[FRmatch]{FR.test}}.}
}
\value{
A list of FR test results:
\item{pmat}{A matrix of p-values for pairwise comparisons. Rows are reference clusters, and columns are query clusters.}
\item{statmat}{A matrix of FR statistics for pairwise comparisons. Rows are reference clusters, and columns are query clusters.}
If \code{return.all = TRUE}, more FR test intermediate results are returned.
}
\description{
This is a user-end function that wraps the steps of matching cell type clusters between two single-cell RNA-seq experiments
(namely, \code{query} and \code{reference}) with clustered expression data and NS-Forest marker genes using Friedman-Rafsky (FR) test.
This function inputs two data objects of the \link[SingleCellExperiment]{SingleCellExperiment} data class,
and outputs the FR statistics, p-values, and optional intermediated restuls from the FR test.
}
\examples{
\dontrun{
data("sce.example")
FRmatch(sce.example, sce.example)
}

}
\seealso{
Visualization of matching results using \code{\link[FRmatch]{plot_FRmatch}}, \code{\link[FRmatch]{plot_bilateral_FRmatch}}.
}
\author{
Yun Zhang, \email{zhangy@jcvi.org};
Brian Aevermann, \email{baeverma@jcvi.org};
Richard Scheuermann, \email{RScheuermann@jcvi.org}.
}
