% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRmatch.R
\name{FRmatch}
\alias{FRmatch}
\title{FR-Match for single cell RNA-seq data}
\usage{
FRmatch(
  sce.query,
  sce.ref,
  filter.size = 5,
  filter.fscore = NULL,
  method = "subsampling",
  subsamp.size = 20,
  subsamp.iter = 1000,
  subsamp.seed = 1,
  numCores = NULL,
  prefix = c("query.", "ref."),
  verbose = 1,
  return.all = FALSE,
  ...
)
}
\arguments{
\item{sce.query}{Data object of the \link[SingleCellExperiment]{SingleCellExperiment} class for query experiment.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{sce.ref}{Data object of the \link[SingleCellExperiment]{SingleCellExperiment} class for reference experiment.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{filter.size, filter.fscore}{Filtering small/poor-quality clusters. Default: \code{filter.size=5}, filter based on the number
of cells per cluster; \code{filter.fscore=NULL}, filter based on the F-beta score associated with the cell cluster if available (numeric value).}

\item{method}{Methods for the FR test. Default: \code{method="subsampling"} is to iteratively subsample equal number of cells (i.e. subsample size)
from the query and reference clusters, and then perform the FR test. Option: \code{method="none"} is the FR test with no modification.}

\item{subsamp.size, subsamp.iter, subsamp.seed}{Iterative subsampling size, number of iterations, and random seed for iterations. YMMV.}

\item{numCores}{Number of cores for parallel computing.
Default: \code{NULL}, use the maximum number of cores detected by \code{\link[parallel]{detectCores}} if not specified (an integer).}

\item{prefix}{Prefix names for query and reference clusters. Default: \code{prefix=c("query.", "ref.")}.}

\item{verbose}{Numeric value indicating levels of details to be printed. Default: \code{1}, only print major steps.
If \code{0}, no verbose; if \code{2}, print all.}

\item{return.all}{Logical variable indicating if to return all results (such as runs, etc.). Default: \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[FRmatch]{FRtest}}, including \code{use.cosine}.}
}
\value{
A list of:
\item{settings}{Record of customized parameter settings specified in the function.}
\item{pmat}{A matrix of p-values. Rows are reference clusters, and columns are query clusters.}
\item{statmat}{A matrix of FR statistics. Rows are reference clusters, and columns are query clusters.}
If \code{return.all = TRUE}, more intermediate results are returned.
}
\description{
This is a user-end function that wraps up the steps of matching cell type clusters between two single cell RNA-seq experiments
(namely, \code{query} and \code{reference}) with \emph{clustered} expression data and \emph{informative} marker genes
using the Friedman-Rafsky (FR) statistical test.
This function inputs two data objects of the \link[SingleCellExperiment]{SingleCellExperiment} class,
and outputs the FR statistics, p-values, and optionally, all intermediate restuls from the FR test.
}
\examples{
\dontrun{
data("sce.example")
FRmatch(sce.example, sce.example)
}

}
\seealso{
Visualization of matching results using \code{\link[FRmatch]{plot_FRmatch}}, \code{\link[FRmatch]{plot_bi_FRmatch}}.
}
\author{
Yun Zhang, \email{zhangy@jcvi.org};
Brian Aevermann, \email{baeverma@jcvi.org};
Richard Scheuermann, \email{RScheuermann@jcvi.org}.
}
