% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRmatch_cell2cluster.R
\name{FRmatch_cell2cluster}
\alias{FRmatch_cell2cluster}
\title{Cell-to-cluster extension of \code{FRmatch}}
\usage{
FRmatch_cell2cluster(
  sce.query,
  sce.ref,
  filter.size = 10,
  filter.fscore = NULL,
  subsamp.size = 10,
  subsamp.iter = 1001,
  subsamp.seed = 1,
  numCores = NULL,
  prefix = c("query.", "ref."),
  verbose = 1,
  ...
)
}
\arguments{
\item{sce.query}{Data object of the \link[SingleCellExperiment]{SingleCellExperiment} class for query experiment.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{sce.ref}{Data object of the \link[SingleCellExperiment]{SingleCellExperiment} class for reference experiment.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{filter.size, filter.fscore}{Filtering small/poor-quality clusters. Default: \code{filter.size=10}, filter based on the number
of cells per cluster; \code{filter.fscore=NULL}, filter based on the f-score associated with the cell cluster if available (numeric value).}

\item{subsamp.size, subsamp.iter, subsamp.seed}{Cluster size, number of iterations, and random seed.
Default: \code{10, 1001, 1}, respectively.}

\item{numCores}{Number of cores for parallel computing.
Default: \code{NULL}, use the maximum number of cores detected by \code{\link[parallel]{detectCores}} if not specified (an integer).}

\item{prefix}{Prefix names for query and reference clusters. Default: \code{prefix=c("query.", "ref.")}.}

\item{verbose}{Numeric value indicating levels of details to be printed. Default: \code{1}, only print major steps.
If \code{0}, no verbose; if \code{2}, print all.}

\item{...}{Additional arguments passed to \code{\link[FRmatch]{FR.test}}.}
}
\value{
A data frame with columns:
\item{cell}{Query cell ID.}
\item{cluster}{Cluster membership of query cells.}
\item{match.cell2cluster}{Matched reference cluster for the query cell by \code{FRmatch_cell2cluster}.}
\item{rmax.cell2cluster}{Row maximum of \code{pmat.cell2cluster} (see below).}
And concatenated by the columns from the matrix:
\item{pmat.cell2cluster}{Cell-by-cluster (a.k.a. query cell by reference cluster) matrix of p-values by \code{FRmatch_cell2cluster}.}
}
\description{
This function is an extension of the original \code{\link[FRmatch]{FRmatch}} to assign each query cell with a reference cluster label.
Please see Details for the extension.
}
\details{
Apply \code{FRmatch} with its iterative subsampling scheme, which is a bootstrap-like approach to randomly select a smaller set of cells
from a query cluster and quantify the confidence score of the selected cells belonging to certain reference cell type
using the p-value outputted from \code{FRmatch} (i.e. a larger p-value indicates higher probability of a match, and vice versa).

Assign the cluster-level p-value to each selected query cell, and updated the assigned p-value if the query cell is reselected
from the iterative procedure and assigned a higher p-value. The output from the cell-to-cluster extension is a cell-by-cluster
(a.k.a. query cell by reference cluster) matrix of p-values.
}
\examples{
\dontrun{
data("sce.example")
FRmatch_cell2cluster(sce.example, sce.example)
}

}
\author{
Yun Zhang, \email{zhangy@jcvi.org};
Brian Aevermann, \email{baeverma@jcvi.org};
Richard Scheuermann, \email{RScheuermann@jcvi.org}.
}
