% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRmatch_cell2cluster.R
\name{FRmatch_cell2cluster}
\alias{FRmatch_cell2cluster}
\title{FR-Match cell-to-cluster matching}
\usage{
FRmatch_cell2cluster(
  sce.query,
  sce.ref,
  filter.size = 10,
  filter.fscore = NULL,
  subsamp.size = 5,
  subsamp.iter = 2000,
  subsamp.seed = 1,
  numCores = NULL,
  prefix = c("query.", "ref."),
  verbose = 1,
  ...
)
}
\arguments{
\item{sce.query}{Data object of the \link[SingleCellExperiment]{SingleCellExperiment} class for query experiment.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{sce.ref}{Data object of the \link[SingleCellExperiment]{SingleCellExperiment} class for reference experiment.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{filter.size, filter.fscore}{Filtering small/poor-quality clusters. Default: \code{filter.size=10}, filter based on the number
of cells per cluster; \code{filter.fscore=NULL}, filter based on the F-beta score associated with the cell cluster if available (numeric value).}

\item{subsamp.size, subsamp.iter, subsamp.seed}{Iterative subsampling size, number of iterations, and random seed for iterations. YMMV.}

\item{numCores}{Number of cores for parallel computing.
Default: \code{NULL}, use the maximum number of cores detected by \code{\link[parallel]{detectCores}} if not specified (an integer).}

\item{prefix}{Prefix names for query and reference clusters. Default: \code{prefix=c("query.", "ref.")}.}

\item{verbose}{Numeric value indicating levels of details to be printed. Default: \code{1}, only print major steps.
If \code{0}, no verbose; if \code{2}, print all.}

\item{...}{Additional arguments passed to \code{\link[FRmatch]{FRtest}}, including \code{use.cosine}.}
}
\value{
A list of:
\item{settings}{Record of customized parameter settings specified in the function.}
\item{pmat}{A cell-by-cluster (a.k.a. query cell by reference cluster) matrix of p-values retained from the iterative procedure.}
\item{cell2cluster}{A data frame of cell-to-cluster matches summarized from the \code{pmat}.}
Columns in \code{cell2cluster} are:
\item{query.cell}{Query cell ID.}
\item{query.cluster}{Cluster membership of query cells.}
\item{match}{Matched reference cluster for the query cell.}
\item{score}{Confidence score of \code{match.cell2cluster}, which is the maximum value of the corresponding row in \code{pmat}.}
}
\description{
This function is an implementation of the FR-Match pipeline that matches each query cell to a reference cluster.
}
\details{
This implementation is \code{FRmatch} with an iterative subsampling scheme, which is a bootstrap-like approach to randomly select a smaller
set of cells from a query cluster and quantify the confidence score of the selected cells belonging to certain reference cell type
using the p-value outputted from \code{FRtest} (i.e. a larger p-value indicates higher probability of a match, and vice versa).

This function assigns the cluster-level p-value to each selected query cell, and updates the assigned p-value if the query cell is reselected
from the iterative procedure and assigned a higher p-value. The output from this implementation includes a cell-by-cluster
(a.k.a. query cell by reference cluster) matrix of p-values.
}
\examples{
\dontrun{
data("sce.example")
FRmatch_cell2cluster(sce.example, sce.example)
}

}
\author{
Yun Zhang, \email{zhangy@jcvi.org};
Brian Aevermann, \email{baeverma@jcvi.org};
Richard Scheuermann, \email{RScheuermann@jcvi.org}.
}
