% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRmatch_cell2cluster.R
\name{FRmatch_cell2cluster}
\alias{FRmatch_cell2cluster}
\title{Cell-to-cluster cell type matching algorithm for single-cell RNA-seq data}
\usage{
FRmatch_cell2cluster(sce.query, sce.ref, filter.size = 10,
  filter.fscore = NULL, subsamp.size = 10, subsamp.iter = 1001,
  subsamp.seed = 1, numCores = NULL, prefix = c("query.", "ref."),
  verbose = 1, ...)
}
\arguments{
\item{sce.query}{Data object of the \link[SingleCellExperiment]{SingleCellExperiment} class for query experiment.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{sce.ref}{Data object of the \link[SingleCellExperiment]{SingleCellExperiment} class for reference experiment.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{filter.size, filter.fscore}{Filtering small/poor-quality clusters. Default: \code{filter.size=10}, filter based on the number
of cells per cluster; \code{filter.fscore=NULL}, filter based on the f-score associated with the cell cluster if available (numeric value).}

\item{subsamp.size, subsamp.iter, subsamp.seed}{Cluster size, number of iterations, and random seed.
Default: \code{10, 1001, 1}, respectively.}

\item{numCores}{Number of cores for parallel computing.
Default: \code{NULL}, use the maximum number of cores detected by \code{\link[parallel]{detectCores}} if not specified (an integer).}

\item{prefix}{Prefix names for query and reference clusters. Default: \code{prefix=c("query.", "ref.")}.}

\item{verbose}{Numeric value indicating levels of details to be printed. Default: \code{1}, only print major steps.
If \code{0}, no verbose; if \code{2}, print all.}

\item{...}{Additional arguments passed to \code{\link[FRmatch]{FR.test}}.}

\item{imputation}{INACTIVE. Logical variable indicating if to impute expression zero values for the reference experiment. Default: \code{FALSE}.
See details in \code{\link[FRmatch]{impute_dropout}}.}
}
\value{
A matrix with columns:
\item{match.cell2cluster}{}
}
\description{
This is a user-end function that wraps up the steps of matching cell type clusters between two single-cell RNA-seq experiments
(namely, \code{query} and \code{reference}) with \emph{clustered} expression data and \emph{informative} marker genes
using the Friedman-Rafsky (FR) statistical test.
This function inputs two data objects of the \link[SingleCellExperiment]{SingleCellExperiment} class,
and outputs the FR statistics, p-values, and optionally, all intermediate restuls from the FR test.
}
\examples{
\dontrun{
data("sce.example")
FRmatch_cell2cluster(sce.example, sce.example)
}

}
\seealso{
Visualization of matching results using \code{\link[FRmatch]{plot_FRmatch}}, \code{\link[FRmatch]{plot_bilateral_FRmatch}}.
}
\author{
Yun Zhang, \email{zhangy@jcvi.org};
Brian Aevermann, \email{baeverma@jcvi.org};
Richard Scheuermann, \email{RScheuermann@jcvi.org}.
}
