% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_object.R
\name{make_data_object}
\alias{make_data_object}
\title{Make data object}
\usage{
make_data_object(
  dat,
  tab,
  markers,
  cluster_marker_info = NULL,
  fscores = NULL,
  cluster_order = NULL
)
}
\arguments{
\item{dat}{Cell-by-gene expression data in a data frame or equivalent format. First column is cell names, followed by columns for each gene.}

\item{tab}{Cluster membership of cells in a data frame or equivalent format. First column is cell names, and second column is cluster labels.}

\item{markers}{A vector of marker genes.}

\item{cluster_marker_info}{Optionally, a data frame of maker genes for each cluster. See details in \code{\link[FRmatch]{sce.example}}.}

\item{fscores}{Optionally, a data frame of F-scores for each cluster. See details in \code{\link[FRmatch]{sce.example}}.}

\item{cluster_order}{Optionally, a vector of ordered cluster names. See details in \code{\link[FRmatch]{sce.example}}.}
}
\value{
A data object of the \link[SingleCellExperiment]{SingleCellExperiment} class
}
\description{
This function helps to create the data object that can be used in the \code{\link[FRmatch]{FRmatch}} algorithm with input information.
}
