% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{Normalization function for FR-Match analysis.}
\usage{
normalization(sce.object, scale = TRUE, norm.by = "mean", impute.zero = FALSE)
}
\arguments{
\item{sce.object}{Data object.}

\item{scale}{Boolean variable indicating if to scale the PER GENE expression to the range [0,1]. Defualt: \code{TRUE}.}

\item{norm.by}{Normalization options PER CLUSTER. Choose from \code{"softmax", "mean", "median", "none"}.
If \code{"softmax"} (default), use the \code{\link[mclust]{softmax}} function to normalize all values globally.
Afterwards, each cell is rescaled to unit length (i.e., L2 normalization).
If \code{"mean", "median"}, each gene is weighted by the row-wise statistic to de-noise lowly expressed values.
Afterwards, all values are rescaled to [0,1].
If \code{"none"}, do nothing.}
}
\value{
A data object with normalized expression data.
}
\description{
Normalization function for FR-Match analysis.
}
