% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bilateral_FRmatch.R
\name{plot_bilateral_FRmatch}
\alias{plot_bilateral_FRmatch}
\title{Bilateral matching plot of FR-Match results}
\usage{
plot_bilateral_FRmatch(rst.FRmatch.E1toE2, rst.FRmatch.E2toE1,
  name.E1 = "E1", name.E2 = "E2", p.adj.method = "BY",
  sig.level = 0.05, reorder = TRUE, return.value = FALSE,
  cellwidth = 10, cellheight = 10, main = NULL, filename = NA, ...)
}
\arguments{
\item{rst.FRmatch.E1toE2, rst.FRmatch.E2toE1}{The \code{\link[FRmath]{FRmatch}} outputs.}

\item{name.E1, name.E2}{Customized names for E1 and E2. Default: \code{"E1"} and \code{"E2"}, respectively.}

\item{p.adj.method}{See \code{\link[FRmath]{plot_FRmatch}}.}

\item{sig.level}{See \code{\link[FRmath]{plot_FRmatch}}.}

\item{reorder}{See \code{\link[FRmath]{plot_FRmatch}}.}

\item{return.value}{Logical variable indicating if to return plotted values. Default: \code{FALSE}.}

\item{cellwidth, cellheight, main, filename, ...}{Plotting parameters passed to \code{\link[pheatmap]{pheatmap}}.}
}
\value{
If \code{return.value = TRUE}, a matrix of matching results. 2 = two-way match, 1 = one-way match, and 0 = no match.
}
\description{
This function combines two sets of reciprocal \code{\link[FRmath]{FRmatch}} outputs (i.e. experiment 1 (hereinafter, E1) as the query
dataset mapping to experiment 2 (hereinafter, E2) as the reference dataset, and E2 mapping to E1), and plots the two-way matches,
one-way matches, and no matches determined by FR-Match.
}
\seealso{
\code{\link[FRmatch]{plot_FRmatch}}.
}
