% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FRmatch.R
\name{plot_FRmatch}
\alias{plot_FRmatch}
\title{Plotting function for FRmatch results}
\usage{
plot_FRmatch(rst.FRmatch, type = "matches", sig.level = 0.05,
  cellwidth = 10, cellheight = 10, ...)
}
\arguments{
\item{rst.FRmatch}{The output retults from the \code{\link[FRmath]{FRmatch}} function.}

\item{type}{Character variable that specifies which element from the output list to be ploted.
Default: \code{"matches"}, which returns the recommended matches found by FRmatch p-values thresholded at \code{sig.level}.
Other options: \code{"pmat", "statmat"}.}

\item{sig.level}{Numeric variable that sets the significance level for FRmatch p-values, above which is a match.
Default value: \code{0.05}.}

\item{cellwidth, cellheight, ...}{Additional plotting parameters passed to \code{\link[pheatmap]{pheatmap}}.}
}
\description{
This function facilitates visualization of \code{FRmatch} results.
}
\seealso{
\link[pheatmap]{pheatmap}.
}
