% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FRmatch.R
\name{plot_FRmatch}
\alias{plot_FRmatch}
\title{Plot FR-Match results}
\usage{
plot_FRmatch(rst.FRmatch, type = "matches", p.adj.method = "BY",
  sig.level = 0.05, reorder = TRUE, return.value = FALSE,
  cellwidth = 10, cellheight = 10, main = NULL, filename = NA, ...)
}
\arguments{
\item{rst.FRmatch}{The \code{\link[FRmath]{FRmatch}} output.}

\item{type}{If \code{type="matches"}, it plots the final matching results.
If \code{type=="padj"}, it plots the distribution of adjusted p-values.}

\item{p.adj.method}{P-value adjustment method for multiple comparison correction. Default: \code{"BY"}.
Please see \code{\link[stats]{p.adjust.methods}}.}

\item{sig.level}{Numeric variable that specifies the significance level of adjusted p-value, above which is a match.
Default value: \code{0.05}.}

\item{reorder}{Logical variable indicating if to reorder the columns of the heatmap of matching results, which may have
better interpretability (aligning matches in the diagonal). Default: \code{TRUE}.}

\item{return.value}{Logical variable indicating if to return plotted values. Default: \code{FALSE}.}

\item{cellwidth, cellheight, main, ...}{Plotting parameters passed to \code{\link[pheatmap]{pheatmap}}.}
}
\value{
If \code{return.value = TRUE}, a binary matrix of matching results or a matrix of adjusted p-values.
}
\description{
This function takes in the \code{\link[FRmath]{FRmatch}} output and provides plots of the results. If \code{type="matches"},
it plots the final matching results. If \code{type=="padj"}, it plots the distribution of adjusted p-values.
}
\seealso{
\code{\link[FRmatch]{plot_bilateral_FRmatch}}.
}
