% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FRmatch.R
\name{plot_FRmatch}
\alias{plot_FRmatch}
\title{Plotting function for FR-Match results}
\usage{
plot_FRmatch(
  rst.FRmatch,
  type = "matches",
  p.adj.method = "BY",
  sig.level = 0.05,
  refine = TRUE,
  reorder = TRUE,
  ignore.unassigned = FALSE,
  return.value = FALSE,
  main = NULL,
  cellwidth = 10,
  cellheight = 10,
  filename = NA,
  ...
)
}
\arguments{
\item{rst.FRmatch}{The \code{\link[FRmatch]{FRmatch}} output.}

\item{type}{If \code{type="matches"} (default), it plots the one-way matches.
If \code{type="padj"}, it plots the distribution of adjusted p-values.}

\item{p.adj.method}{P-value adjustment method for multiple hypothesis testing correction. Default: \code{"BY"}.
For more options, please see \code{\link[stats]{p.adjust.methods}}.}

\item{sig.level}{Numeric variable that specifies the significance level of adjusted p-value. A MATCH is >\code{sig.level}.
Default: \code{0.05}.}

\item{refine}{Boolean variable. Default: \code{TRUE}. If \code{TRUE}, matches are refined based on \eqn{\delta > 0.2},
where \eqn{\delta} is the difference between the maximum \code{padj} and other non-significant (i.e. > \code{sig.level}) \code{padj} for each query cluster.}

\item{reorder}{Boolean variable indicating if to reorder the columns so that matches are aligned along the diagonal.
It improves the interpretability of the one-way match plot. Default: \code{TRUE}.}

\item{ignore.unassigned}{Boolean variable indicating if to skip the columns of unassigned query clusters
in the \code{type="matches"} plot. Default: \code{FALSE}. If \code{TRUE}, number of ignored columns reported in the unassigned row.}

\item{return.value}{Boolean variable indicating if to return the plotted values. Default: \code{FALSE}.}

\item{main}{Plot title.}

\item{cellwidth, cellheight, filename, ...}{Plotting parameters passed to \code{\link[pheatmat]{pheatmap}}.}
}
\value{
If \code{return.value = TRUE}, a matrix of one-way matching values 1 = match, and 0 = no match, or a matrix of adjusted p-values.
}
\description{
This function takes in the \code{\link[FRmatch]{FRmatch}} output and generates plots for the matching results.
}
\seealso{
\code{\link[FRmatch]{plot_bi_FRmatch}}.
}
