% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FRmatch_cell2cluster.R
\name{plot_FRmatch_cell2cluster}
\alias{plot_FRmatch_cell2cluster}
\title{Plotting function for FR-Match cell2cluster results}
\usage{
plot_FRmatch_cell2cluster(
  rst.cell2cluster,
  type = "match",
  p.adj.method = "BH",
  sig.level = 0.1,
  reorder = TRUE,
  return.value = FALSE,
  main = NULL,
  filename = NA,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{rst.cell2cluster}{The \code{\link[FRmatch]{FRmatch_cell2cluster}} output.}

\item{type}{If \code{type="match"} (default), it plots the proportion of cells in the query cluster matched to the reference cluster.
If \code{type=="padj"}, it plots the distribution of adjusted p-values.}

\item{p.adj.method}{P-value adjustment method for multiple hypothesis testing correction. Default: \code{"BH"}.
For more options, please see \code{\link[stats]{p.adjust.methods}}.}

\item{sig.level}{Numeric variable that specifies the significance level of adjusted p-value. A MATCH is >\code{sig.level}. Default: \code{0.1}.}

\item{reorder}{Boolean variable indicating if to reorder the columns so that matches are aligned along the diagonal. Default: \code{TRUE}.}

\item{return.value}{Boolean variable indicating if to return the plotted values. Default: \code{FALSE}.}

\item{main}{Plot title.}

\item{filename, width, height, units}{Plotting parameters passed to \code{\link[ggplot2]{ggsave}}.}
}
\value{
If \code{return.value = TRUE}, a matrix of \code{plotted.values}, and \code{pmat.adj} and \code{cell2cluster.adj} after adjustment.
}
\description{
This function takes in the \code{\link[FRmatch]{FRmatch_cell2cluster}} output and generates plots for the matching results.
}
