% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nonzero.R
\name{plot_nonzero}
\alias{plot_nonzero}
\title{Plot per-cluster percentage of non-zero expression for markers}
\usage{
plot_nonzero(sce.object, return.plot = TRUE, return.value = FALSE,
  cellwidth = 15, cellheight = 10, main = NULL, ...)
}
\arguments{
\item{sce.object}{A \code{SingleCellExperiment} object customized with necessary information for \code{FRmatch}.
See details in \code{\link[FRmatch]{sce.example}}.}

\item{return.plot}{Logical variable indicating if to return a plot. Default: \code{TRUE}.}

\item{return.value}{Logical variable indicating if to return the values correponding to the plot. Default: \code{FALSE}.}

\item{cellwidth, cellheight, main, ...}{Additional plotting parameters passed to \code{\link[pheatmap]{pheatmap}}.}
}
\value{
Optionally, a numeric matrix corresponding to the values on the plot.
}
\description{
A function that calculates and plots percentage of non-zero expression for each marker gene per cluster
for a \code{SingleCellExperiment} object customized with cluster membership and marker gene information.
}
\examples{
\dontrun{
data("sce.example")
plot_nonzero(sce.example)
}

}
\seealso{
The \link[SingleCellExperiment]{SingleCellExperiment} class.
}
