% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nonzero.R
\name{plot_nonzero}
\alias{plot_nonzero}
\title{Non-zero expression plot}
\usage{
plot_nonzero(sce.object, return.plot = TRUE, return.value = FALSE,
  cellwidth = 15, cellheight = 10, main = NULL, ...)
}
\arguments{
\item{sce.object}{A \code{FRmatch} input data object. See example in \code{\link[FRmatch]{sce.example}}.}

\item{return.plot}{Logical variable indicating if to return the plot. Default: \code{TRUE}.}

\item{return.value}{Logical variable indicating if to return the plotted values. Default: \code{FALSE}.}

\item{cellwidth, cellheight, main, ...}{Plotting parameters passed to \code{\link[pheatmap]{pheatmap}}.}
}
\value{
if \code{return.value = TRUE}, a matrix of plotted values.
}
\description{
A function that calculates and plots "\% expressed per marker gene per cluster" for the \code{FRmatch} input data object.
\% = number of cells that express the marker gene in the cluster / cluster size.
}
\examples{
\dontrun{
data("sce.example")
plot_nonzero(sce.example)
}

}
