% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce.example}
\alias{sce.example}
\title{FR-Match input data}
\format{Please follow the names and data formats of this example when constructing your own data.
\describe{
\item{\code{assays}}{Data assay \code{logcounts} is the gene expression matrix to be used.}
\item{\code{rownames}}{Gene names.}
\item{\code{rowData}}{Column \code{marker_gene} is a boolean vector indicating if a gene is a marker gene (1) or not (0).}
\item{\code{colnames}}{Cell names.}
\item{\code{colData}}{Column \code{cluster_membership} is a character vector of cell cluster membership.}
\item{\code{metadata}}{List of 3.
\itemize{
  \item \code{cluster_marker_info} is a data frame of maker genes for each cluster.
  \item \code{fsocres} is a data frame of F-scores for each cluster.
  \item \code{cluster_order} is a vector of ordered cluster names, which will be used in the final output and plots.
}}
}}
\usage{
sce.example
}
\description{
This page describes the \link[SingleCellExperiment]{SingleCellExperiment} data object that can be used
as input data to \code{\link[FRmatch]{FRmatch}}. This example dataset contains 16497 genes and 865 cells.
}
\seealso{
The \link[SingleCellExperiment]{SingleCellExperiment} data class.
}
\keyword{datasets}
